Working Prototype
Finished remaking the file imports from importing a concatenated list to a list of objects. This will make it a lot easier to add functionalities in the future. I was then able to remake the list of duplicate invoices. The newly added feature was removing the list of duplicate invoices from the XML file.